requirements.txt
Django==3.2.12
django-celery-beat==1.5.0
celery==4.4.1
alpha-vantage==2.3.1
djangorestframework==3.13.1